// src/auth/enums/permission.enum.ts
export enum Permission {
  // File Operations
  FILE_UPLOAD = 'PERMISSION:FILE_UPLOAD',
  FILE_SAVE = 'PERMISSION:FILE_SAVE',
  FILE_DELETE = 'PERMISSION:FILE_DELETE',
  FILE_EDIT = 'PERMISSION:FILE_EDIT',

  // Folder Operations
  FOLDER_DELETE = 'PERMISSION:FOLDER_DELETE',
  FOLDER_CREATE = 'PERMISSION:FOLDER_CREATE',

  // Keyword Operations
  KEYWORD_CHECK_UNUSED = 'PERMISSION:KEYWORD_CHECK_UNUSED',
  KEYWORD_CREATE = 'PERMISSION:KEYWORD_CREATE',
  KEYWORD_DELETE = 'PERMISSION:KEYWORD_DELETE',

  // WebSocket Operations (Admin only)
  WEBSOCKET_PREVIEWS_SYNC = 'PERMISSION:WEBSOCKET_PREVIEWS_SYNC',
  WEBSOCKET_PREVIEWS_CREATE = 'PERMISSION:WEBSOCKET_PREVIEWS_CREATE',
  WEBSOCKET_EXIF_GET = 'PERMISSION:WEBSOCKET_EXIF_GET',

  // Testing Operations (Admin only)
  TEST_MATCHING_NUMBER_OF_FILES = 'PERMISSION:TEST_MATCHING_NUMBER_OF_FILES',
}